requirements (1).txt
Flask==2.3.2
flask-cors==3.0.10
requests==2.31.0
python-dotenv==1.0.0
gunicorn==20.1.0
